requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.75.0
    # via ynab-amazon-itemize (pyproject.toml)
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   starlette
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via functions-framework
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
flask==3.1.2
    # via functions-framework
functions-framework==3.10.0
    # via ynab-amazon-itemize (pyproject.toml)
google-api-core==2.29.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-secret-manager
google-api-python-client==2.187.0
    # via ynab-amazon-itemize (pyproject.toml)
google-auth==2.47.0
    # via
    #   ynab-amazon-itemize (pyproject.toml)
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-secret-manager
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-auth-oauthlib==1.2.2
    # via ynab-amazon-itemize (pyproject.toml)
google-cloud-core==2.5.0
    # via google-cloud-firestore
google-cloud-firestore==2.22.0
    # via ynab-amazon-itemize (pyproject.toml)
google-cloud-secret-manager==2.26.0
    # via ynab-amazon-itemize (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-secret-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
gunicorn==23.0.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via anthropic
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.12.0
    # via anthropic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-cloud-secret-manager
protobuf==6.33.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via anthropic
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.1
    # via httplib2
python-dateutil==2.9.0.post0
    # via ynab-amazon-itemize (pyproject.toml)
python-dotenv==1.2.1
    # via ynab-amazon-itemize (pyproject.toml)
requests==2.32.5
    # via
    #   ynab-amazon-itemize (pyproject.toml)
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anthropic
starlette==0.50.0
    # via functions-framework
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.5
    # via
    #   flask
    #   functions-framework
